fix(locale): Use applicationContext for getting resource configuration
Instead of accessing system resources directly, we can use the application
context to fetch the resource configurations and subsequently the locale
used in the device. Not only is it safer to do this instead of accessing
system resources directly, it also helps with the client's use-case that
they can modify and set a particular locale for the application.