feat(moneymanager-ui): Add polling with timeout to statistics fetching
A workaround with polling to handle cases such as when a user first goes into the Finance Overview the ingestion of transactions into Tink might not be done and we don't want to show an empty state then. A timeout after 15 seconds is added so that we don't remain in the loading state forever and we can assume
that sufficient time has passed for the transactions to be ingested.